Scoring system
I started working on the scoring system that is required to assess marginality. In addition, infinitival subjects are not marginal due to parsing problems, but because they make bad topics; hence I had to include one piece of discourse computation to factor this variable correctly. I am otherwise still assessing whether the model agrees with intuition for the infinitival corpus, which has turned out to be rather tricky.