Support for partial pro-drop
Added support for partial pro-drop (and partial inflection) and removed MOR. Inflectional features accepted by any given head must be given explicitly, if not, they are not accepted. Antecedent requirements are stated in the output (computed at the outer edge of LF) but not implemented.

Accepted phi feature A and incoming phi feature B match if A can be matched with the beginning of B. For example, INFL=PHI matches with all features that begin with PHI. This is a regex system (but not yet implemented thus). You can therefore give verb general accetance criteria (INFL=PHI, e.g. Italian) or very narrow (INFL=PHI:NUM:SG, INFL=PHI:PER:3, as in English).