Major improvements and changes
Extraposition algorithm was not able to handle clitic data (was based on wrong assumptions), which required that it was ordered differently, which in turn implied several other architectural changes that had to be made.  This update implements these changes but does not yet handle all the clitic data correctly (requires additional changes). The update also prepares for the ecentual context-mechanism, in that each process now have access to the parser's "sentence_context" variable that contains language, discoures informations and so on.