Improvements to quantitative datatake
This version added more support for controlling and configuring the internal parameters of the model. It also improved the multistudy functionality. The output now contains more detailed timing information (in ms). Most of the required functionality for quantitative analysis are done; the algorithm no longer completely satisfies observational adequacy, however, which will be fixed later.