Syntactic working memory first implementation
This version implements a notion of active syntactic working memory based on psycholinguistic literature and some other concerns. It will also add a log file that contains the derivations in a simple and more readable format. Finally, this version contains all simulation experiments for the performance paper. The model has not been checked for observational and descriptive adequacy.