Refined DP/NP structure to prepare BT. This is not fully tested version.
BT requires a more fine-grained DP/NP structure, in essence to handle the situation in which the input argument has no overt D element. To this end, the phi-layer was added. This is the most elegant solution I can currently think of. This version works but requires testing because it was a major change.