Recursive assignment function (skeleton)
Recursive assignment function was added which generates all possible assignments for any given sentence (assignment = possible mapping between QND expressions and entities in the ontology). Nontrivial part is expressions such as 'two men' and 'every men'. This version works only in the context of finite semantics (generalization to infinite domains will be added later). This is a skeleton function for testing purposes that assumes every QND denotes every entity in the discourse inventory.