Simplifications to several modules (e.g., LF, tail tests)
There will now be a series of simplifications and unifications to the codebase. This will be implemented in several stages to keep the operation feasible. OA will be maintained for each step. The idea is to push the codebase towards the abstract "shape" required by the data -- and thus elimiate coding specific noise and imperfections.