Refactoring and comments (Study 16b, 20, OA-2)
This is a prewar update before I will leave the project at least for a while. Added comments and some refactoring. The EPP and clitics studies (16b, 20) are combined here, but there is a separate subfolder for study 20 from where it should be run. The documentation is inadequate (lags behind) and will be updated once some new study goes into print.